Package ctypes 0.24.0

ComplexL

Cstubs Operations for generating C bindings stubs.

Cstubs_analysis

Cstubs_c_language

Cstubs_emit_c

Cstubs_errors

Cstubs_generate_c

Cstubs_generate_ml

Cstubs_internals

Cstubs_inverted Operations for exposing OCaml code as C libraries.

Cstubs_public_name

Cstubs_structs

Ctypes The core ctypes module.

Ctypes_bigarray

Ctypes_bigarray_stubs

Ctypes_coerce

Ctypes_memory

Ctypes_memory_stubs

Ctypes_path

Ctypes_primitive_types

Ctypes_primitives

Ctypes_ptr

Ctypes_roots_stubs

Ctypes_static

Ctypes_std_view_stubs

Ctypes_std_views

Ctypes_structs

Ctypes_structs_computed Structs and unions whose layouts are computed from the sizes and alignment requirements of the constituent field types.

Ctypes_top

Ctypes_type_printing

Ctypes_types

Ctypes_value_printing

Ctypes_value_printing_stubs

LDouble

PosixTypes Some POSIX types.

Package info

authors

Jeremy Yallop

changes-files

CHANGES.md

depends

conf-fts

conf-pkg-config

dune

dune-configurator

integers

ocaml

odoc

ounit2

homepage

https://github.com/yallop/ocaml-ctypes

issues

https://github.com/yallop/ocaml-ctypes/issues

license

MIT

license-files

LICENSE

maintainers

Jeremy Yallop <yallop@gmail.com>

online-doc

https://yallop.github.io/ocaml-ctypes/

readme-files

README.md

repo

git+https://github.com/yallop/ocaml-ctypes.git

tags

org:mirage

org:ocamllabs

version

0.24.0